Problem: Random test timeouts
Solution: Switch back to the `HttpProvider`.

After switching to IPC geth connection, some of
the tests started to randomly timeout.
Serializing some of the parallel blockchain transactions
seemed to lower the probability of timeouts, but
they would only disappear when using the `HttpProvider`,
which is sometimes a lot slower.

Also disable whisper, since it's not used at the moment.

The test chain is listening on port 9545 until this issue
is resolved.